+++ kanso/cms/wrappers/Visit.php
@@ -19,10 +19,7 @@ class Visit extends Wrapper
 	use MagicArrayAccessTrait;
 
     /**
-     * Saves the current visitor.
-     *
-     * @access public
-     * @return bool
+     * {@inheritdoc}
      */
     public function save(): bool
     {
@@ -60,6 +57,52 @@ class Visit extends Wrapper
         if (isset($this->data['id']))
         {
             $this->SQL->DELETE_FROM('crm_visits')->WHERE('id', '=', $this->data['id'])->QUERY();
+
+            return true;
+        }
+
+        return false;
+    }
+
+    /**
+     * Returns all visit actions.
+     *
+     * @access public
+     * @return array
+     */
+    public function actions(): array
+    {
+        if (isset($this->data['id']))
+        {
+            return $this->SQL->SELECT('*')->FROM('crm_visit_actions')->WHERE('visit_id', '=', $this->data['id'])->FIND_ALL();
+        }
+
+        return [];
+    }
+
+    /**
+     * Adds a new action.
+     *
+     * @access public
+     * @param  string $action      The action type
+     * @param  string $description The action description
+     * @return bool
+     */
+    public function addAction(string $action, string $description): bool
+    {
+        if (isset($this->data['id']))
+        {
+            $row =
+            [
+                'visit_id'           => $this->data['id'],
+                'visitor_id'         => $this->data['visitor_id'],
+                'action_name'        => $action,
+                'action_description' => $description,
+                'page'               => $this->data['page'],
+                'date'               => time(),
+            ];
+
+            $this->SQL->INSERT_INTO('crm_visit_actions')->VALUES($row)->QUERY();
 
             return true;
         }